requirements.txt
langchain>=0.0.220
langchain-core>=0.0.220
langchain-perplexity>=0.0.220
langchain-community>=0.0.25
langgraph>=0.0.15
pandas>=2.0.0
numpy>=1.25.0
streamlit>=1.24.0
altair>=5.0.1
requests>=2.28.0
python-dateutil>=2.8.2
tqdm>=4.65.0
python-dotenv>=1.0.0
langchain-huggingface
transformers
huggingface-hub
pyyaml
graphviz
